init: leaderBoardApi
